WIP (refactor): move telegram handlers and auth logic into separate modules
