+++ config.py
@@ -1,0 +1,1 @@
+# TODO вынести сюда константы из bot_state
